# agentcms.web/specs/001-admin-portal/contracts/sites.openapi.yaml
openapi: 3.0.1
info:
  title: AgentCMS Admin API - Sites
  description: Site management endpoints for the AgentCMS Admin Portal
  version: "1.0"

paths:
  /v1/sites:
    get:
      summary: List all sites
      operationId: listSites
      tags:
        - Sites
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SiteDto'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    
    post:
      summary: Create a new site
      operationId: createSite
      tags:
        - Sites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteDto'
      responses:
        '201':
          description: Site created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteDto'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Server error

  /v1/sites/{id}:
    get:
      summary: Get a specific site
      operationId: getSite
      tags:
        - Sites
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteDto'
        '404':
          description: Site not found
        '500':
          description: Server error
    
    put:
      summary: Update a site
      operationId: updateSite
      tags:
        - Sites
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSiteDto'
      responses:
        '200':
          description: Site updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteDto'
        '400':
          description: Validation error
        '404':
          description: Site not found
        '500':
          description: Server error
    
    delete:
      summary: Delete a site
      operationId: deleteSite
      tags:
        - Sites
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Site deleted successfully
        '404':
          description: Site not found
        '500':
          description: Server error

components:
  schemas:
    SiteDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 255
          nullable: true
      required:
        - id
        - name
    
    CreateSiteDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Site name (required)
        description:
          type: string
          maxLength: 255
          nullable: true
          description: Optional site description
      required:
        - name
    
    UpdateSiteDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Site name (required)
        description:
          type: string
          maxLength: 255
          nullable: true
          description: Optional site description
      required:
        - name
    
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: true